Modify grade class to handle grade given a score
Add Course and Score attribute in addition to Grades in the Grade
Class. Grade class stores course code, score and grade which is handled
by the parseGrade function. Grade is not to be inputted by user.

Modify parseGrade method to parse scores into
corresponding grades and not converting enum grade to string format.
Create GradeUtil helper function and move parseGrade method from Grade into
GradeUtil Class.

Create Course class and Score class to ensure that course and score are
in the correct format.